+++ kits/braze/braze-43/example/example-kotlin/src/main/kotlin/com/mparticle/kits/braze/braze43/example/kotlin/ExampleApplication.kt
@@ -1,4 +1,4 @@
-package com.mparticle.kits.braze.braze42.example.kotlin
+package com.mparticle.kits.braze.braze43.example.kotlin
 
 import android.app.Application
 import com.mparticle.MPEvent